+++ relm4-examples/examples/factory.rs
@@ -1,6 +1,6 @@
 use gtk::glib::Sender;
 use gtk::prelude::{BoxExt, ButtonExt, GtkWindowExt, OrientableExt};
-use relm4::factory::{Factory, FactoryPrototype, FactoryVec};
+use relm4::factory::{FactoryPrototype, FactoryVec};
 use relm4::{send, AppUpdate, Model, RelmApp, WidgetPlus, Widgets};
 
 #[derive(Debug)]